+++ app-framework/tests/deos_council_board.rs
@@ -72,7 +72,7 @@ fn board() -> (AppCipherclerk, EmbeddedExecutor, DeosCell) {
             AuthRequired::Either,
             Effect::SetField {
                 cell: proposal,
-                index: STATUS_SLOT,
+                index: STATUS_SLOT as u64,
                 value: fe(RESOLVED),
             },
         ),
