+++ dregg-storage-templates/src/programmable_queue.rs
@@ -421,12 +421,12 @@ pub fn build_enqueue_action(
     let effects = vec![
         Effect::SetField {
             cell: queue_cell,
-            index: HEAD_SEQ_SLOT as usize,
+            index: HEAD_SEQ_SLOT as u64,
             value: new_head,
         },
         Effect::SetField {
             cell: queue_cell,
-            index: RING_ROOT_SLOT as usize,
+            index: RING_ROOT_SLOT as u64,
             value: new_ring_root,
         },
         Effect::EmitEvent {
@@ -451,7 +451,7 @@ pub fn build_dequeue_action(
     let effects = vec![
         Effect::SetField {
             cell: queue_cell,
-            index: TAIL_SEQ_SLOT as usize,
+            index: TAIL_SEQ_SLOT as u64,
             value: new_tail,
         },
         Effect::EmitEvent {
@@ -477,7 +477,7 @@ pub fn build_grant_sender_action(
     let effects = vec![
         Effect::SetField {
             cell: queue_cell,
-            index: SENDER_SET_ROOT_SLOT as usize,
+            index: SENDER_SET_ROOT_SLOT as u64,
             value: new_sender_set_root,
         },
         Effect::EmitEvent {
